requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
supabase
google-generativeai
python-dotenv
PyJWT
feedparser
requests
